requirements.txt
pandas
numpy
tqdm
google-cloud-bigquery
db-dtypes
pandas-gbq